BUG FIXES: Frame::buildMessageFromFrames clears the message before building the message from Frames

+++ include/hexabitz/Service.h
@@ -79,7 +79,7 @@ private:
 
 private:
 	std::shared_ptr<ProxyModule> owner_;
-	uint8_t num_modules_;
+	int num_modules_;
 	ModulesInfo info_;
 };
